RegisterController랑 TestController를  UserController로 병합
