+++ application.log
@@ -1,5 +1,18 @@
-2025-11-13 07:58:27 INFO  [SpringApplicationShutdownHook] o.s.b.w.e.tomcat.GracefulShutdown - Commencing graceful shutdown. Waiting for active requests to complete
-2025-11-13 07:58:27 INFO  [tomcat-shutdown] o.s.b.w.e.tomcat.GracefulShutdown - Graceful shutdown complete
-2025-11-13 07:58:27 INFO  [SpringApplicationShutdownHook] o.s.o.j.LocalContainerEntityManagerFactoryBean - Closing JPA EntityManagerFactory for persistence unit 'default'
-2025-11-13 07:58:27 INFO  [SpringApplicationShutdownHook] com.zaxxer.hikari.HikariDataSource - HikariPool-1 - Shutdown initiated...
-2025-11-13 07:58:27 INFO  [SpringApplicationShutdownHook] com.zaxxer.hikari.HikariDataSource - HikariPool-1 - Shutdown completed.
+2025-11-13 09:42:00 WARN  [restartedMain] o.s.b.a.o.j.JpaBaseConfiguration$JpaWebConfiguration - spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning
+2025-11-13 09:42:00 DEBUG [restartedMain] o.s.w.s.m.m.a.RequestMappingHandlerMapping - 62 mappings in 'requestMappingHandlerMapping'
+2025-11-13 09:42:00 DEBUG [restartedMain] o.s.w.s.h.SimpleUrlHandlerMapping - Patterns [/webjars/**, /ui/**, /swagger-ui*/*swagger-initializer.js, /swagger-ui*/**] in 'resourceHandlerMapping'
+2025-11-13 09:42:00 INFO  [restartedMain] o.s.b.a.e.web.EndpointLinksResolver - Exposing 1 endpoint beneath base path '/actuator'
+2025-11-13 09:42:01 DEBUG [restartedMain] o.s.w.s.m.m.a.RequestMappingHandlerAdapter - ControllerAdvice beans: 0 @ModelAttribute, 0 @InitBinder, 1 RequestBodyAdvice, 1 ResponseBodyAdvice
+2025-11-13 09:42:01 DEBUG [restartedMain] o.s.w.s.m.m.a.ExceptionHandlerExceptionResolver - ControllerAdvice beans: 2 @ExceptionHandler, 1 ResponseBodyAdvice
+2025-11-13 09:42:02 INFO  [restartedMain] o.a.coyote.http11.Http11NioProtocol - Starting ProtocolHandler ["http-nio-0.0.0.0-8080"]
+2025-11-13 09:42:02 INFO  [restartedMain] o.s.b.w.e.tomcat.TomcatWebServer - Tomcat started on port 8080 (http) with context path '/'
+2025-11-13 09:42:02 INFO  [restartedMain] h.cluebackend.ClueBackendApplication - Started ClueBackendApplication in 8.699 seconds (process running for 10.17)
+2025-11-13 09:42:03 INFO  [RMI TCP Connection(3)-127.0.2.3] o.a.c.c.C.[Tomcat].[localhost].[/] - Initializing Spring DispatcherServlet 'dispatcherServlet'
+2025-11-13 09:42:03 INFO  [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Initializing Servlet 'dispatcherServlet'
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Detected StandardServletMultipartResolver
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Detected AcceptHeaderLocaleResolver
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Detected FixedThemeResolver
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Detected org.springframework.web.servlet.view.DefaultRequestToViewNameTranslator@1acb4179
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Detected org.springframework.web.servlet.support.SessionFlashMapManager@3ecad06a
+2025-11-13 09:42:03 DEBUG [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - enableLoggingRequestDetails='false': request parameters and headers will be masked to prevent unsafe logging of potentially sensitive data
+2025-11-13 09:42:03 INFO  [RMI TCP Connection(3)-127.0.2.3] o.s.web.servlet.DispatcherServlet - Completed initialization in 5 ms
